is_esml -> in_master_block WIP FIXME
+        # FIXME test what is the actual behavior for older games
flag or extension

Now that the esl dust settled a bit it's clear we should separately
handle the load order and light plugin esl behavior. This commit is
about the load order part moving the load order handling in the game
module where it belongs and exposing needed functionality via
load_order.py.